added just so much

+++ source/scenes/game.h
@@ -108,6 +108,6 @@ class game: public Gamemode
 		void checkLines(int(blocks)[240]);
 		bool checkRow(int(blocks)[10]);
 		void clearRow(int(blocks)[240], int y);
-		void drawCubes(int position[],double scale, int x, int y, int size, int width,bool threed=false, glm::vec3 rotation={0,0,0});
+		void drawCubes(int position[],float scale, float x, float y, int size, int width,bool threed=false, glm::vec3 rotation={0,0,0});
 };
